Add basics for ZoneFactory
Zones can only be created by the owner of the
contract. I intend to allow mobs to be created
in a particular ZONE, and that mob must be of
the appropriate level. Users can then "explore"
a given zone to see a list of mobs that they can
attack.

This time I tried using the expectRevert test
helper. You can specify the revert reason or
use expectRevert.unspecified() to check against
ANY exception, though I think it's best to specify
a reason to avoid false positives.
- https://docs.openzeppelin.com/learn/writing-automated-tests