Add basics for Zones and Mobs
This section gave me a lot of trouble;
I started off wanting to get a random mob
in ZoneFactory from a pool of mobs in MobFactory,
but I couldn't figure out a simple implementation
where ZoneFactory could access data and types in
MobFactory. For now, I've decided just to pass
the name / description for the ZoneMob directly
into the function rather than look it up in
the Mob array. In the next step I will make
this createZoneMob internal, so I will be able
to control the inputs myself to ensure ZoneMobs
created will match whatever is in the bestiary /
mobs array in MobFactory.

In looking into all this I also stumbled on a great
reply posted on StackExchange:
- https://ethereum.stackexchange.com/questions/13167/are-there-well-solved-and-simple-storage-patterns-for-solidity
This showed me that it's common to use a getEntityCount()
function instead of trying to reference the entity_array.length
property in the javascript code.

I'm still not sure how I want to handle ZoneIn, etc.
I'm considering that I want users to adventure to a
certain zone and be presented with a number of mobs
that land in the appropriate level range, though
I'm not sure how best to handle this, and whether or
not these ZoneMobs even need to be stored in data.
My hunch is that it isn't necessary; it's enough
to maintain the bestiary in MobFactory.

Finally, I stumbled on the idea of using proxy
contracts to maintain data storage and delegate
contracts to hold application logic:
https://blog.gnosis.pm/solidity-delegateproxy-contracts-e09957d0f201
https://medium.com/quillhash/how-to-write-upgradable-smart-contracts-in-solidity-d8f1b95a0e9a