Add Character Can Attack Mobs, Gain XP and Level Up
I rearranged the instantiation of the Character Contract
to make use of the ZoneFactory so I could access the
zone mob details. Though I only need the mob level,
this seemed safer than simply passing the mob level
to the contract. There is no level checking in the
attack() method to safeguard against any abuse.
Perhaps doing that would be cheaper, gas-wise,
than fetching and loading the mob into memory, though.

The battle / experience gained handling is rather simple,
but it is fine for this simple project / demo; I've never
built a game before, but I imagine programming a more
sophisticated battle system is quite the task.

I'll have to do some refactoring here to get more
safety / clarity around Character experience values.
I KNOW that each level is just 250 xp, and that battles
will never yield more than this, and my calcs reflect this
understanding, but I don't think this would be clear to
another dev, and there are probably numerous ways this
could break. Now that things work and tests are in place
I will work to make this more secure and robust in terms
of error handling if I can.

Noting my use of zoneFactory.getZoneMob() it is annoying
that I can't just use an automatic getter to get the mob struct;
perhaps this is because the zoneMobs array is nested in
a map. Regardless, I can't seem to state the zoneFactory.Mob
as a type in the CharacterFactory, anyway.

As I work with Solidity I'm regularly frustrated that
there are so many simple things I can't do.
Hopefully the language will continue to improve.
I have read that returning structs is something in the works.

Finally, it seems many things are possible through the
use of various getters - eg: getEntityCount. I will
have to spend more time getting used to these patterns:
https://ethereum.stackexchange.com/questions/13167/are-there-well-solved-and-simple-storage-patterns-for-solidity