WIP Add Go server w/ basic admin page + webpack
Intro

This is just to stub in what will eventually
b/c a slightly-modularized front end.

I like to keep the JS to a minimum. This is
partially b/c I'm pretty sick of JS frameworks
(I've used them a lot), but also because I
like the speed of a site that doesn't use a
lot of JS.

Bootstrap Template

I swiped this from another project using
a Bootstrap template. I'll modify this
a lot as I work, so I'm not too concerned with
getting all the styles correct, removing
unused code, etc. at this point;
I'm more concerned with just arranging my build
system so I can write small JS modules and load
them on the proper pages.
You can see that each module will get a reference
to the Contract objects.

Svelte?

I may decide to use Svelte with this project.
It is something I've wanted to use for a while,
and it will be a nice change of pace from Vue / React.

Webpack Build

The scripts used to build a dev and prod bundle
are in package.json, of course. A dev build is
important for debugging.

Go Backend

This isn't too robust at the moment. I'll likely
make the webserver portion a bit better in the future.