Add storage helper to manage user data
This is a simple helper to keep track
of the user account being used to interact
with the app. I'm not sure if this is even
the correct way to go about this, but I'm
anticipating having to keep track of
a user's "selected" character for when that
user wants to adventure and battle mobs, etc.

I wanted to record last_visited timestamps,
too, as I think that would be good for the
user to quickly see when they last used a
character. None of this is wired up in the
UI, yet, but this is the groundwork I think
I'll need to start experimenting with this.

I'll send the userAccount object through to
each module, too, so it will be easy to save
and retrieve user data throughout the app.