Add EDIT Zone Support
Now the Owner of the contract (admin)
can edit a zone's name and description.
The Owner can also edit level range IF
there aren't any mobs added to the zone,
yet.

Thoughts:

I think the API would be more clear if I
had two separate methods:

1 - editZoneNameAndDescription
2 - editZoneLevelRange

Pros:

As it is now, the change to levels will
be ignored, and without the UI check,
a user wouldn't even know why. In fact,
it would be better to return an Error
when the level change shouldn't be processed.
We can only do this if the change to the levels
is explicitely requested by the user.

Cons:

The downside is that with different
methods, the UI needs to be changed.
This isn't something I'm ready to do
now.