WIP Add Zone Admin Page w/ Create Zone Mob
This is just to get the page going.
I still have to add the ability to edit
zone name / description, and perhaps support
to delete zone mobs.
I would also like to show a thumbnail and
details when a particulr mob type is selected
in the form.

Ultimately, it would be cool if I could
add special loot / bonuses to the zone mob
that could be randomnly given to a
user if they defeat this mob while adventuring.