Add selectCharacter() to User Storage helper
Remove updateLastVisited() as that wasn't
necessary with save()
This was really a pointless method I was
just using to test this anyway.
Perhaps I would like to track last_used
or last_battle for each character,
though I'm not sure if that would be
useful without storing that data on the
blockchain.