Add select character to character screen
A user can now select a character and
store that selection in local storage for
the account activated when the page loaded.
I believe I will eventually have to start
refreshing pages when a different account
is selected in MetaMask, etc. but I'm not
sure how best to do that yet.